Create classAutoPath.ahk
Hello,

Here's an auto-complete AutoHotkey v2 Class.  

Up to 20 completion suggestions will be shown as user inputs partial file/folder paths.  
/ (divide) key cycles through suggestions (will be highlighted/selected)
Enter key accepts suggestions (unhighlights/deselects this suggestion).  If nothing is highlighted, Enter will submit the result and close the GUI.
< (less than) character enables folder only mode
> (greater than) character enables file only mode
" (double quote) charater indicates provided file/folder path pattern is literal and * (asterisk) will not be appended automatically
* and ? are supported

The overall result is similiar to Windows' Run command with support of wildcards (?/*) without the need to use mouse/arrow keys.

Example usage:
[code]a := new classAutoPath(A_WinDir, "w500")
while true {
	if a.ready || a.error
		break
	sleep 100
}

if a.error
	MsgBox("error")
if a.ready
	MsgBox(a.path)[/code]

This creates an classAutoPath object with the default folder set to A_WinDir and a width of 500.

Properties: 
.ready is set to 1 once user has submit a file/folder path (invalid paths are not checked).  Default is 0.
.error is set to 1 if user exits GUI prior to submission (0 otherwise)
.path is the submitted path as string